fix(577-aru): address codex review — loader/generator durability + hash overflow guard
Codex review of PR #590 (all verified before fixing):

- HIGH: src/ml/synthetic/loaders/batch_loader.py TABLE_COLUMNS["triggers"] omitted
  action_taken + control_group_flag, and _load_batch filters each DataFrame to that
  whitelist (`df[columns_to_load]`) before upsert — so a fresh synthetic load would
  silently DROP both columns and the registry query's `WHERE control_group_flag IS
  NOT NULL` would find empty arms -> fail-loud. Added both to the whitelist +
  regression-lock test (test_triggers_whitelist_carries_action_rate_uplift_columns).
- MEDIUM: trigger_generator._generate_standalone_triggers (the non-linked path) did
  not emit the arm/action columns. Mirrored the arm-conditioned logic there too (the
  loader of record uses the linked _generate_trigger_record path, already done, but
  parity matters). mypy: replaced np.where(..., None) with a typed list comprehension.
- LOW: migration 051 abs(hashtext(...)) can overflow at INT_MIN — cast ::bigint before
  abs in all 3 hash sites. Re-applied to supabase-db; realized uplift UNCHANGED
  (0.2751341, value-preserving for all non-INT_MIN hashes).

The faithful target was already correct (migration 051 reseeded all 4356 live rows);
these close the fresh-regeneration durability gap. 89 kpi+loader tests green; ruff +
mypy clean.

+++ database/migrations/051_kpi_577_action_rate_uplift.sql
@@ -96,7 +96,7 @@ ALTER TABLE public.triggers ADD COLUMN IF NOT EXISTS control_group_flag boolean;
 --     Independent of the outcome and of any covariate => a legitimate randomized
 --     assignment. Live-verified: n_control=1222, n_treatment=3134 (0.2805 control).
 UPDATE public.triggers
-   SET control_group_flag = (abs(hashtext(trigger_id || 'arm')) % 100) < 28;
+   SET control_group_flag = (abs(hashtext(trigger_id || 'arm')::bigint) % 100) < 28;
 
 -- (C) Arm-conditioned action_taken reseed over ALL 4356 rows (idempotent /
 --     re-runnable). Treatment P(action present)=0.38, control P=0.30 => a real
@@ -107,10 +107,10 @@ UPDATE public.triggers
 --     only seed the data, they are NOT the returned value.
 UPDATE public.triggers
    SET action_taken = CASE
-       WHEN (abs(hashtext(trigger_id || 'act')) % 100)
+       WHEN (abs(hashtext(trigger_id || 'act')::bigint) % 100)
             < (CASE WHEN control_group_flag THEN 30 ELSE 38 END)
        THEN (ARRAY['called_patient', 'scheduled_visit', 'sent_info'])[
-                1 + (abs(hashtext(trigger_id || 'verb')) % 3)]
+                1 + (abs(hashtext(trigger_id || 'verb')::bigint) % 3)]
        ELSE NULL END;
 
 -- (D) Register the read-only WS2-TR-003 statement (allowlist; executed only via
